+++ spring-boot-email-core/src/main/java/it/ozimov/springboot/templating/mail/model/Email.java
@@ -44,6 +44,7 @@ public interface Email {
         return "";
     }
 
+    //FIXME Release 0.4.0 will fix this bug to have Collection<EmailAttachment>
     Collection<EmailAttachmentImpl> getAttachments();
 
     Charset getEncoding();