+++ spring-boot-email-core/src/main/java/it/ozimov/springboot/templating/mail/model/impl/EmailAttachmentImpl.java
@@ -40,7 +40,6 @@ import static java.util.Optional.ofNullable;
 /**
  * Plain text email.
  *
- * @author rtrunfio
  */
 @NoArgsConstructor
 @RequiredArgsConstructor