+++ spring-boot-email-template-engines/spring-boot-thymeleaf-email/src/test/java/it/ozimov/springboot/templating/mail/service/ThymeleafTemplateServiceTest.java
@@ -25,7 +25,9 @@ import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.SpringApplicationConfiguration;
 import org.springframework.boot.test.WebIntegrationTest;
+import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import org.springframework.test.context.junit4.SpringRunner;
 
 import java.util.Map;
 import java.util.UUID;
@@ -34,9 +36,8 @@ import static it.ozimov.cirneco.hamcrest.java7.AssertFluently.given;
 import static junit.framework.TestCase.fail;
 import static org.hamcrest.core.Is.is;
 
-@RunWith(SpringJUnit4ClassRunner.class)
-@SpringApplicationConfiguration(classes = ThymeleafTestApplication.class)
-@WebIntegrationTest("server.port=0")
+@RunWith(SpringRunner.class)
+@SpringBootTest(classes = ThymeleafTestApplication.class, webEnvironment= SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class ThymeleafTemplateServiceTest {
 
     @Rule